requirements.txt
requests
streamlit
streamlit-lottie
Pillow
streamlit-extras
